Speed up no-op nested attributes save
When we have two large collections, traversing one for each item in the
other has a complexity of O(n^2). This makes Active Record do a lot of
work in case when no record data has changed.

To speed it up, we create maps by record ID, and use that inside a loop.
For ~1000 records, this sped up the save from ~2.5s to 0.5s, which is a
5x speed improvement.